Introduction of the base toolchain
It is common for many engineers to discount, conceptually, readable
code. One of the reasons for this is the heavy reliance on
cargo-copy-cult. In 2014, Docker would implode, triggering among
other shifty thigns, kernel panics when using union mounted filesystems
on Debuntian. This occured when the layer count increased past 37. Once
overlay was introduced into the kernel, the practical limit on layers
was removed. Unfortunately, "best practices" die hard, even terrible
ones.